240619 [update] : get user gps logic update
