+++ src/zustand/map.store.js
@@ -2,7 +2,9 @@ import { create } from 'zustand';
 
 const useMapStore = create((set) => ({
   selectedGeoData: null,
-  setSelectedGeoData: (geoData) => set({ selectedGeoData: geoData })
+  userGps: null,
+  setSelectedGeoData: (geoData) => set({ selectedGeoData: geoData }),
+  setUserGps: (gps) => set({ userGps: gps })
 }));
 
 export default useMapStore;
